Update meetings display
If agenda available for next meeting, provide link to it.

Only get next meetings from past meetings.

Previously it was assumed that there would always be minutes or
presentation materials for the last meeting, but this is not always the
case, so add a case to display nothing about last meeting if neither of
these are available.